refactor: tighten preview runtime module boundaries
