+++ src/VersoBlueprint/Commands/preview-runtime-base.mjs
@@ -78,11 +78,116 @@
       .replaceAll("'", "&#39;");
   }
 
+  export function normalizePreviewMode(rawMode, fallback) {
+    const defaultMode = fallback === "hover" || fallback === "pinned" ? fallback : "hover";
+    const mode = String(rawMode || "").trim().toLowerCase();
+    if (mode === "hover") return "hover";
+    if (mode === "pinned") return "pinned";
+    return defaultMode;
+  }
+
+  export function normalizePreviewPlacement(rawPlacement, fallback) {
+    const defaultPlacement =
+      fallback === "anchored" || fallback === "docked" ? fallback : "anchored";
+    const placement = String(rawPlacement || "").trim().toLowerCase();
+    if (placement === "anchored") return "anchored";
+    if (placement === "docked") return "docked";
+    return defaultPlacement;
+  }
+
+  export function readPanelBehavior(panel, defaults) {
+    const defaultMode = normalizePreviewMode(defaults && defaults.mode, "hover");
+    const defaultPlacement = normalizePreviewPlacement(defaults && defaults.placement, "anchored");
+    if (!(panel instanceof Element)) {
+      return {
+        mode: defaultMode,
+        placement: defaultPlacement,
+        isPinned: defaultMode === "pinned",
+        isHover: defaultMode === "hover",
+        isAnchored: defaultPlacement === "anchored",
+        isDocked: defaultPlacement === "docked"
+      };
+    }
+    const rawMode = (panel.getAttribute("data-bp-preview-mode") || "").trim();
+    const rawPlacement = (panel.getAttribute("data-bp-preview-placement") || "").trim();
+    const mode = normalizePreviewMode(rawMode, defaultMode);
+    const placement = normalizePreviewPlacement(rawPlacement, defaultPlacement);
+    return {
+      mode: mode,
+      placement: placement,
+      isPinned: mode === "pinned",
+      isHover: mode === "hover",
+      isAnchored: placement === "anchored",
+      isDocked: placement === "docked"
+    };
+  }
+
+  export function normalizePanelBehavior(panel, defaults, nextBehavior) {
+    const fallback = readPanelBehavior(panel, defaults);
+    if (!nextBehavior || typeof nextBehavior !== "object") return fallback;
+    const mode = normalizePreviewMode(nextBehavior.mode, fallback.mode);
+    const placement = normalizePreviewPlacement(nextBehavior.placement, fallback.placement);
+    return {
+      mode: mode,
+      placement: placement,
+      isPinned: mode === "pinned",
+      isHover: mode === "hover",
+      isAnchored: placement === "anchored",
+      isDocked: placement === "docked"
+    };
+  }
+
+  export function readStringOption(options, name, fallback) {
+    return options && typeof options[name] === "string" && options[name].length > 0
+      ? options[name]
+      : fallback;
+  }
+
+  export function readObjectOption(options, name, fallback) {
+    return options && options[name] && typeof options[name] === "object"
+      ? options[name]
+      : fallback;
+  }
+
+  export function readNumberOption(options, name, fallback) {
+    return options && Number.isFinite(options[name])
+      ? options[name]
+      : fallback;
+  }
+
+  export function readFunctionOption(options, name, fallback) {
+    return options && typeof options[name] === "function"
+      ? options[name]
+      : fallback;
+  }
+
+  export function readElementOption(options, name, fallback) {
+    return options && options[name] instanceof Element
+      ? options[name]
+      : fallback;
+  }
+
+  export function readRootOption(options, name, fallback) {
+    return options && (options[name] instanceof Element || options[name] instanceof Document)
+      ? options[name]
+      : fallback;
+  }
+
   export const previewRuntimeBase = {
     previewDebugEnabled,
     previewDebugLabel,
     previewDebug,
     collectPreviewTemplates,
     readHtml,
-    escapeHtml
+    escapeHtml,
+    normalizePreviewMode,
+    normalizePreviewPlacement,
+    readPanelBehavior,
+    normalizePanelBehavior,
+    readStringOption,
+    readObjectOption,
+    readNumberOption,
+    readFunctionOption,
+    readElementOption,
+    readRootOption
   };
