+++ src/VersoBlueprint/Commands/preview-runtime-surface.mjs
@@ -1,4 +1,4 @@
-import { escapeHtml } from "./preview-runtime-base.mjs";
+import { escapeHtml, normalizePanelBehavior, normalizePreviewMode, normalizePreviewPlacement, readElementOption, readFunctionOption, readNumberOption, readObjectOption, readStringOption } from "./preview-runtime-base.mjs";
 import { renderHtmlInto, resolveBlueprintPreview } from "./preview-runtime-render.mjs";
 import { bindCloseOnce, bindDismissHandlers, bindPanelRepositioner, bindPreviewTriggers, positionAnchoredPanel, readAnchorRect, shouldKeepOpen } from "./preview-runtime-lifecycle.mjs";
 
@@ -6,65 +6,6 @@ import { bindCloseOnce, bindDismissHandlers, bindPanelRepositioner, bindPreviewT
   //
   // These helpers own panel slots, behavior state, content updates,
   // and runtime diagnostic markup for bundled clients.
-
-  export function normalizePreviewMode(rawMode, fallback) {
-    const defaultMode = fallback === "hover" || fallback === "pinned" ? fallback : "hover";
-    const mode = String(rawMode || "").trim().toLowerCase();
-    if (mode === "hover") return "hover";
-    if (mode === "pinned") return "pinned";
-    return defaultMode;
-  }
-
-  export function normalizePreviewPlacement(rawPlacement, fallback) {
-    const defaultPlacement =
-      fallback === "anchored" || fallback === "docked" ? fallback : "anchored";
-    const placement = String(rawPlacement || "").trim().toLowerCase();
-    if (placement === "anchored") return "anchored";
-    if (placement === "docked") return "docked";
-    return defaultPlacement;
-  }
-
-  export function readPanelBehavior(panel, defaults) {
-    const defaultMode = normalizePreviewMode(defaults && defaults.mode, "hover");
-    const defaultPlacement = normalizePreviewPlacement(defaults && defaults.placement, "anchored");
-    if (!(panel instanceof Element)) {
-      return {
-        mode: defaultMode,
-        placement: defaultPlacement,
-        isPinned: defaultMode === "pinned",
-        isHover: defaultMode === "hover",
-        isAnchored: defaultPlacement === "anchored",
-        isDocked: defaultPlacement === "docked"
-      };
-    }
-    const rawMode = (panel.getAttribute("data-bp-preview-mode") || "").trim();
-    const rawPlacement = (panel.getAttribute("data-bp-preview-placement") || "").trim();
-    const mode = normalizePreviewMode(rawMode, defaultMode);
-    const placement = normalizePreviewPlacement(rawPlacement, defaultPlacement);
-    return {
-      mode: mode,
-      placement: placement,
-      isPinned: mode === "pinned",
-      isHover: mode === "hover",
-      isAnchored: placement === "anchored",
-      isDocked: placement === "docked"
-    };
-  }
-
-  export function normalizePanelBehavior(panel, defaults, nextBehavior) {
-    const fallback = readPanelBehavior(panel, defaults);
-    if (!nextBehavior || typeof nextBehavior !== "object") return fallback;
-    const mode = normalizePreviewMode(nextBehavior.mode, fallback.mode);
-    const placement = normalizePreviewPlacement(nextBehavior.placement, fallback.placement);
-    return {
-      mode: mode,
-      placement: placement,
-      isPinned: mode === "pinned",
-      isHover: mode === "hover",
-      isAnchored: placement === "anchored",
-      isDocked: placement === "docked"
-    };
-  }
 
   export function resetPanelPosition(panel) {
     if (!(panel instanceof Element)) return;
@@ -450,44 +391,6 @@ import { bindCloseOnce, bindDismissHandlers, bindPanelRepositioner, bindPreviewT
     }
   }
 
-  // Option readers.
-
-  export function readStringOption(options, name, fallback) {
-    return options && typeof options[name] === "string" && options[name].length > 0
-      ? options[name]
-      : fallback;
-  }
-
-  export function readObjectOption(options, name, fallback) {
-    return options && options[name] && typeof options[name] === "object"
-      ? options[name]
-      : fallback;
-  }
-
-  export function readNumberOption(options, name, fallback) {
-    return options && Number.isFinite(options[name])
-      ? options[name]
-      : fallback;
-  }
-
-  export function readFunctionOption(options, name, fallback) {
-    return options && typeof options[name] === "function"
-      ? options[name]
-      : fallback;
-  }
-
-  export function readElementOption(options, name, fallback) {
-    return options && options[name] instanceof Element
-      ? options[name]
-      : fallback;
-  }
-
-  export function readRootOption(options, name, fallback) {
-    return options && (options[name] instanceof Element || options[name] instanceof Document)
-      ? options[name]
-      : fallback;
-  }
-
   export function createPreviewPanel(options) {
     const opts = options && typeof options === "object" ? options : {};
     const panel = document.createElement("aside");
@@ -607,10 +510,6 @@ import { bindCloseOnce, bindDismissHandlers, bindPanelRepositioner, bindPreviewT
   }
 
   export const previewRuntimeSurface = {
-    normalizePreviewMode,
-    normalizePreviewPlacement,
-    readPanelBehavior,
-    normalizePanelBehavior,
     resetPanelPosition,
     hidePreviewSurfaces,
     configureCloseButton,
@@ -620,12 +519,6 @@ import { bindCloseOnce, bindDismissHandlers, bindPanelRepositioner, bindPreviewT
     createPreviewSurface,
     renderPreviewIntoSurface,
     resolvePreviewHtml,
-    readStringOption,
-    readObjectOption,
-    readNumberOption,
-    readFunctionOption,
-    readElementOption,
-    readRootOption,
     createPreviewPanel,
     previewMessageHtml,
     setPreviewHeaderLink
